requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.2.3
Django==3.1
django-filter==2.3.0
django-hashedfilenamestorage==2.4.1
django-tinymce==3.0.2
djangorestframework==3.11.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
gprof2dot==2019.11.30
greenlet==0.4.17
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
locust==1.2.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mysqlclient==2.0.1
Pillow==7.2.0
psutil==5.7.2
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.1.0
